RAG-MedGemma/pyproject.toml
[project]
name = "a8p5-hands-on-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.9.0",
    "bitsandbytes>=0.46.1",
    "gradio>=5.38.2",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-huggingface>=0.3.1",
    "pillow>=11.3.0",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "torch>=2.7.1",
    "transformers>=4.53.3",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true